+++ CHANGELOG.rst
@@ -18,7 +18,7 @@ Change Log
   - Added extract_file_from_zip to zip_utils.
   - Added http_utils with download function.
   - Added get_app_specific_directory, get_os_name, and get_cpu_architecture_name to misc_utils.
-  - Added normalize_file_path to file_utils.
+  - Added normalize_file_path and are_files_equal to file_utils.
   - Added create_temporary_file_name and remove_temporary_file tmpfile_utils.
   - Minor fix to misc_utils.create_dict (do not create property only if its value is None).